presentation stack removed from new views
